feat: 사 기물의 궁성 규칙 추가
