+++ README.md
@@ -14,13 +14,13 @@
 
 ### 궁
 - [x] `전`, `후`, `좌`, `우`로 한 칸 이동할 수 있다.
-- [ ] `궁성 대각선` 위에 있을 경우, 대각선 방향으로 한 칸 이동할 수 있다.
-- [ ] 이동 목적지가 궁성 내부가 아니면 이동할 수 없다.
+- [x] `궁성 대각선` 위에 있을 경우, 대각선 방향으로 한 칸 이동할 수 있다.
+- [x] 이동 목적지가 궁성 내부가 아니면 이동할 수 없다.
 
 ### 사
 - [x] `전`, `후`, `좌`, `우`로 한 칸 이동할 수 있다.
-- [ ] `궁성 대각선` 위에 있을 경우, 대각선 방향으로 한 칸 이동할 수 있다.
-- [ ] 이동 목적지가 궁성 내부가 아니면 이동할 수 없다.
+- [x] `궁성 대각선` 위에 있을 경우, 대각선 방향으로 한 칸 이동할 수 있다.
+- [x] 이동 목적지가 궁성 내부가 아니면 이동할 수 없다.
 
 ### 졸/병
 - [x] 자신의 진영 기준으로 `전`, `좌`, `우`로 한 칸 이동할 수 있다.
